# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Republic API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /websites:
    get:
      operationId: listWebsites
      tags: [websites]
      summary: List websites
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: isFree
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Website"
  /websites/{id}:
    get:
      operationId: getWebsite
      tags: [websites]
      summary: Get a single website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
        "404":
          description: Not found

  # ── User ─────────────────────────────────────────────────────────────────────
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/membership:
    get:
      operationId: getUserMembership
      tags: [user]
      summary: Get user membership status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "401":
          description: Unauthorized
  /user/membership/checkout:
    post:
      operationId: createCheckoutSession
      tags: [user]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "401":
          description: Unauthorized
...
  /admin/websites:
    get:
      operationId: adminListWebsites
      tags: [admin]
      summary: Admin list all websites
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Website"
    post:
      operationId: adminCreateWebsite
      tags: [admin]
      summary: Create a website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"
  /admin/websites/{id}:
    patch:
      operationId: adminUpdateWebsite
      tags: [admin]
      summary: Update a website
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebsiteUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Website"

  # ── Admin: Users ──────────────────────────────────────────────────────────────
...
  /admin/users/{userId}/membership:
    patch:
      operationId: adminUpdateUserMembership
      tags: [admin]
      summary: Update user membership
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MembershipUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"

  # ── Admin: Support ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Website:
      type: object
      required: [id, name, url, displayDomain, isFree, isActive, categoryId, tabOrder, cardOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        displayDomain:
          type: string
        iconUrl:
          type: ["string", "null"]
        isFree:
          type: boolean
        isActive:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        tabOrder:
          type: integer
        cardOrder:
          type: integer
        canBeTab:
          type: boolean
        canPreload:
          type: boolean
        remembersUrl:
          type: boolean
        customUserAgent:
          type: ["string", "null"]
        injectedCss:
          type: ["string", "null"]
        injectedJs:
          type: ["string", "null"]
        appBannerHideSelectors:
          type: ["string", "null"]
        popupMitigationEnabled:
          type: boolean
        notes:
          type: ["string", "null"]
    WebsiteInput:
      type: object
      required: [name, url, displayDomain]
      properties:
        name:
          type: string
        url:
          type: string
        displayDomain:
          type: string
        iconUrl:
          type: ["string", "null"]
        isFree:
          type: boolean
        isActive:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        tabOrder:
          type: integer
        cardOrder:
          type: integer
        canBeTab:
          type: boolean
        canPreload:
          type: boolean
        remembersUrl:
          type: boolean
        customUserAgent:
          type: ["string", "null"]
        injectedCss:
          type: ["string", "null"]
        injectedJs:
          type: ["string", "null"]
        appBannerHideSelectors:
          type: ["string", "null"]
        popupMitigationEnabled:
          type: boolean
        notes:
          type: ["string", "null"]
    WebsiteUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        displayDomain:
          type: string
        iconUrl:
          type: ["string", "null"]
        isFree:
          type: boolean
        isActive:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        tabOrder:
          type: integer
        cardOrder:
          type: integer
        canBeTab:
          type: boolean
        canPreload:
          type: boolean
        remembersUrl:
          type: boolean
        customUserAgent:
          type: ["string", "null"]
        injectedCss:
          type: ["string", "null"]
        injectedJs:
          type: ["string", "null"]
        appBannerHideSelectors:
          type: ["string", "null"]
        popupMitigationEnabled:
          type: boolean
        notes:
          type: ["string", "null"]

    # ── User ─────────────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, email, displayName, isPro, acceptedTermsAt, acceptedPrivacyAt, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isPro:
          type: boolean
        acceptedTermsAt:
          type: ["string", "null"]
        acceptedPrivacyAt:
          type: ["string", "null"]
        createdAt:
          type: string
        theme:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        theme:
          type: string
        acceptedTermsAt:
          type: ["string", "null"]
        acceptedPrivacyAt:
          type: ["string", "null"]

    # ── Membership ───────────────────────────────────────────────────────────────
    Membership:
      type: object
      required: [userId, plan, status]
      properties:
        userId:
          type: string
        plan:
          type: string
          enum: [free, monthly, annual]
        status:
          type: string
          enum: [active, canceled, past_due, none]
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        currentPeriodEnd:
          type: ["string", "null"]
    MembershipUpdate:
      type: object
      properties:
        plan:
          type: string
        status:
          type: string
    CheckoutInput:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [monthly, annual]
    CheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string
...