# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Republic API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileUpdate {
  displayName?: string;
  theme?: string;
  /** @nullable */
  acceptedTermsAt?: string | null;
  /** @nullable */
  acceptedPrivacyAt?: string | null;
}